+++ src/main/java/com/davies/lab/lander/FormattedModels/ResponseBody/ExternalUse/DOHeadResponseExternal.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 import java.util.Set;
 
 public class DOHeadResponseExternal {
-    private Integer HeadID;
+    private Long HeadID;
     private String SondeName;
     private String SondeNo;
     private String SensorType;
@@ -70,11 +70,11 @@ public class DOHeadResponseExternal {
         LanderID = head.getLanderID().getASDBLanderID();
     }
 
-    public Integer getHeadID() {
+    public Long getHeadID() {
         return HeadID;
     }
 
-    public void setHeadID(Integer headID) {
+    public void setHeadID(Long headID) {
         HeadID = headID;
     }
