+++ src/main/java/com/davies/lab/lander/Models/ProcessedDOData.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 public class ProcessedDOData {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer ID;
+    private Long ID;
     private LocalDateTime Date;
     private Double TempDegC;
     private Double DO;
@@ -33,7 +33,7 @@ public class ProcessedDOData {
         HeadID = headID;
     }
 
-    public ProcessedDOData(Integer ID, LocalDateTime date, Double tempDegC, Double DO, Double weissDoMgL, Double battV, Double GGDOMgL, Double BKDOMgL, ProcessedDOHead headID) {
+    public ProcessedDOData(Long ID, LocalDateTime date, Double tempDegC, Double DO, Double weissDoMgL, Double battV, Double GGDOMgL, Double BKDOMgL, ProcessedDOHead headID) {
         this.ID = ID;
         Date = date;
         TempDegC = tempDegC;
@@ -45,11 +45,11 @@ public class ProcessedDOData {
         HeadID = headID;
     }
 
-    public Integer getID() {
+    public Long getID() {
         return ID;
     }
 
-    public void setID(Integer ID) {
+    public void setID(Long ID) {
         this.ID = ID;
     }
