+++ src/main/java/com/davies/lab/lander/Repositories/ProcessedCTDDataRepository.java
@@ -9,11 +9,11 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProcessedCTDDataRepository extends JpaRepository<ProcessedCTDData, Integer> {
+public interface ProcessedCTDDataRepository extends JpaRepository<ProcessedCTDData, Long> {
     //TODO Add custom SQL Query routes as needed by Frontend
     @Query(value = "SELECT * FROM processedctddata WHERE head_id = :id", nativeQuery = true)
-    List<ProcessedCTDData> findDataByHeadId(@Param("id") Integer id);
+    List<ProcessedCTDData> findDataByHeadId(@Param("id") Long id);
 
     @Query(value = "SELECT * FROM processedctddata WHERE head_id = :id AND date BETWEEN :startDate AND :endDate", nativeQuery = true)
-    List<ProcessedCTDData> findDataByHeadAndDateRange(@Param("id") Integer id, @Param("startDate") String startDate, @Param("endDate") String endDate);
+    List<ProcessedCTDData> findDataByHeadAndDateRange(@Param("id") Long id, @Param("startDate") String startDate, @Param("endDate") String endDate);
 }
